interestlens/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# Google Cloud ADK + Gemini
google-generativeai>=0.4.0

# Google OAuth
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# JWT Authentication
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# Redis
redis>=5.0.0

# Browserbase
httpx>=0.26.0

# Observability
weave>=0.50.0

# Voice Onboarding (Daily + Pipecat)
pipecat-ai>=0.0.43
daily-python>=0.10.1
websockets>=12.0

# Utilities
pydantic>=2.5.0
numpy>=1.26.0
